refactor(mvc_flask): reorganize code structure and improve readability
- Move FlaskMVC class to a separate file `mvc_flask.py` for better organization
- Remove unnecessary imports and unused code from `__init__.py`
- Rename `init_app` method in `FlaskMVC` class to `perform` for better clarity
- Extract configuration logic into separate methods in `FlaskMVC` class for better modularity and readability
- Update method names in `FlaskMVC` class to better reflect their purpose
- Update variable names in `FlaskMVC` class for better clarity
- Update comments and docstrings in `FlaskMVC` class for better understanding
